Disable JavaScript in PageOrientedSpec for features that don't need it
With JavaScript enabled, the JavaScript thread modifies the page and with unlucky timing
the link element needed is either not present when checked for in at-checking, or is stale
when trying to click it. This makes the tests not actually needing the dynamic element
flaky, failing occasionally for content not found or stale elements.